requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
python-telegram-bot==20.3
requests